requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
fastapi[standard]
uvicorn
apache-airflow
pyyaml[srv]
dill
xgboost
mlflow
dvc


-e .